Pulled code and updated to use bootstrap for login. Fixed login/register navigation using bootstrap. Fixed submit user/task for bootstrap to store to mongoDB.
